Github Explorer Repository

Github Explorer Repository

Github Explorer Repository

See also

Methods

Public methods

Ecosystem$new()

Ecosystem$read_USER()

Ecosystem$read_REPO()

Ecosystem$read_PACKAGE()

Ecosystem$read_DEPENDENCY()

Ecosystem$read_FOLLOWING()

Ecosystem$read_SPECTATOR()

Ecosystem$overwrite_USER()

Ecosystem$overwrite_REPO()

Ecosystem$overwrite_PACKAGE()

Ecosystem$overwrite_DEPENDENCY()

Ecosystem$overwrite_FOLLOWING()

Ecosystem$overwrite_SPECTATOR()

Method new()

Instantiate a Repository object

Usage

Ecosystem$new(
  local_path = fs::path_wd("_cache", "tables"),
  remote_path = "s3://tidylab/github.observatory/tables/"
)

Arguments

local_path

(character) A local dir path where files will be stored.

remote_path

(character) A remote dir path on AWS S3 where files will be stored. when you call it again it returns the data from internal meomery.

Method read_USER()

Read the Github information of R Users.

Usage

Ecosystem$read_USER()

Method read_REPO()

Read the Github information of R Packages.

Usage

Ecosystem$read_REPO()

Method read_PACKAGE()

Read the CRAN information of R Packages.

Usage

Ecosystem$read_PACKAGE()

Method read_DEPENDENCY()

Read the CRAN package dependencies of R Packages.

Usage

Ecosystem$read_DEPENDENCY()

Method read_FOLLOWING()

Read who is following who in the R zoo.

Usage

Ecosystem$read_FOLLOWING()

Method read_SPECTATOR()

Read R packages "contributors", "stargazers", and "watchers".

Usage

Ecosystem$read_SPECTATOR()

Method overwrite_USER()

Overwrite the Github information of R Users.

Usage

Ecosystem$overwrite_USER(value)

Arguments

value

(data.frame) Data to store in the Repository.

Method overwrite_REPO()

Overwrite the Github information of R Packages.

Usage

Ecosystem$overwrite_REPO(value)

Arguments

value

(data.frame) Data to store in the Repository.

Method overwrite_PACKAGE()

Overwrite the CRAN information of R Packages.

Usage

Ecosystem$overwrite_PACKAGE(value)

Arguments

value

(data.frame) Data to store in the Repository.

Method overwrite_DEPENDENCY()

Overwrite the CRAN package dependencies of R Packages.

Usage

Ecosystem$overwrite_DEPENDENCY(value)

Arguments

value

(data.frame) Data to store in the Repository.

Method overwrite_FOLLOWING()

Overwrite who is following who in the R zoo.

Usage

Ecosystem$overwrite_FOLLOWING(value)

Arguments

value

(data.frame) Data to store in the Repository.

Method overwrite_SPECTATOR()

Overwrite R packages "contributors", "stargazers", and "watchers".

Usage

Ecosystem$overwrite_SPECTATOR(value)

Arguments

value

(data.frame) Data to store in the Repository.