Database for Queries

Database for Queries

Database for Queries

UserQueryDB

RepoQueryDB

Format

An object of class environment of length 1.

An object of class environment of length 1.

Functions

UserQueryDB: User QueryDB

RepoQueryDB: Repo QueryDB

Methods

Public methods

QueryDB$new()

QueryDB$save()

QueryDB$load()

QueryDB$commit()

QueryDB$rollback()

Method new()

Instantiate a QueryDB object

Usage

QueryDB$new(path, immediate = FALSE)

Arguments

path

(character) A character denoting an existing directory of the Archive for which metadata will be aved and returned.

immediate

(logical) Should queries be committed immediately?

Method save()

Save a query in the database

Usage

QueryDB$save(
  data = list(),
  entity = character(),
  type = character(),
  id = character(),
  alias = character()
)

Arguments

data

(list) The query data.

entity

(character) The entity name. For example "repo" or "user".

type

(character) The query type. For example "overview" or "contributors".

id

(character) The entity id. For example "280924484" or "clintools".

alias

(character) The entity alias name. For example "dplyr" or "hadley".

Method load()

Load queries from database

Usage

QueryDB$load()

Method commit()

Commit entires to database

Usage

QueryDB$commit()

Method rollback()

Rollback changes from the database

Usage

QueryDB$rollback()